scripts/requirement.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.49.0
gunicorn==21.2.0
hubspot-api-client==8.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
mailchimp-transactional==1.0.56
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
pymongo==4.7.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
